+++ source/datastructures/optiblastmatrix.hpp
@@ -19,12 +19,12 @@ public:
     OptiBlastMatrix(string, string, string, bool, double c, int l, float p, bool min);
     ~OptiBlastMatrix(){ }
     
-    vector< set<int> > getBlastOverlap() { return blastOverlap; }
-    string getOverlapName(int); //name from nameMap index
+    vector< set<long long> > getBlastOverlap() { return blastOverlap; }
+    string getOverlapName(long long); //name from nameMap index
     
 protected:
     vector<string> overlapNameMap;
-    vector< set<int> > blastOverlap;  //empty unless reading a blast file.
+    vector< set<long long> > blastOverlap;  //empty unless reading a blast file.
     
     string distFile, namefile, countfile, format;
     bool sim, minWanted;
@@ -32,7 +32,9 @@ protected:
     int length;
     
     int readBlast();
-    int readBlastNames(map<string, int>& nameAssignment);
+    int readBlastNames(map<string, long long>& nameAssignment);
+    
+    
 };
