+++ source/datastructures/optimatrix.h
@@ -21,7 +21,7 @@ public:
     OptiMatrix() : OptiData(0.03) { };
 #endif
     
-    OptiMatrix(vector< set<int> >, vector<string>, vector<string>, double); //closeness, namemap, singleton, cutoff
+    OptiMatrix(vector< set<long long> >, vector<string>, vector<string>, double); //closeness, namemap, singleton, cutoff
     OptiMatrix(string, string, string, string, double, bool); //distfile, dupsFile, dupsFormat, distFormat, cutoff, sim
     ~OptiMatrix(){ }
